apps/discord/requirements.txt
discord.py==2.5.2
httpx==0.28.1
pydantic-settings==2.8.1
python-dotenv==1.1.0
aiohttp==3.13.4
boto3>=1.34.0
